Поменял тип денег с double на int. Сделал тесты для Wallet и частично для AddCalcualator

+++ app/src/main/java/com/rdd/finy/app/views/InfoView.kt
@@ -10,7 +10,7 @@ interface InfoView : MvpView {
 
     fun showControlMoneyDialog(isAddingMoney : Boolean)
 
-    fun updateTotalBalanceState(totalBalance : Double)
+    fun updateTotalBalanceState(totalBalance: Int)
 
     fun expandFAB()
